requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tensorflow
pyyaml>=6.0
matplotlib>=3.7.0
Pillow>=9.5.0
pytest>=7.0.0
